Fix PVQ encoding with multiple tiles
During the encoding loop, td.mb.pvq_q is set to point to the pvq queue
for the current tile (this was done in encode_rd_sb_row but since this
is global to the whole tile I moved it to vp10_encode_tile), the same
needs to be done in the bitstream loop.

To make it harder to use the wrong pvq_q pointer, it is now reset to
NULL after processing a tile.

Note: It would be even better if we didn't need to have a pvq_q pointer
in td.mb, but there is no way to access the TileDataEnc of the current
tile from tokenize_pvq (the index of the current tile or a pointer to
the current tile is not stored anywhere)

Change-Id: Id4b7c3fbdd0f3b97e9dfc703729c117c8a3d60c6